+++ lib/pages/startmenue_seite.dart
@@ -300,7 +300,7 @@ class _StartmenueSeiteState extends State<StartmenueSeite> with RouteAware {
                 const Spacer(),
                 const Center(
                   child: Text(
-                    'Web App 0.9.12 · r336 @ GitHub:',
+                    'Web App 0.9.12 · r336a @ GitHub:',
                     style: TextStyle(fontSize: 13, color: AppFarben.subtilerText),
                   ),
                 ),